qcacld-3.0: Fix Bss scoring algorithm related to channel BW
Scenario: Take a HT40 capable AP and configure it in 20Mhz mode
and then turn on the DUT and scan and connect to a SSID with
multiple BSSIDs present.

Observation: DUT calculates the score of BSSID currently
operating in 20Mhz taking 40Mhz bandwidth which is its max
capablity.

Expectation: DUT should calculate the bandwidth score by
taking 20Mhz current operational bandwidth.

Fix: Check the ht operation IE, and take the AND of the cap
and operation to fill the BSS chan width as 40Mhz.

Change-Id: Ib39714d4d42d5f0484f193e4f4c954d9afecbd99
CRs-Fixed: 2467025